Move competition delete to training page
